+++ tracker_project/tracker/models.py
@@ -10,7 +10,7 @@ class BMI(models.Model):
     weightKilos = models.FloatField(blank = True, null=True)
     
 class Weight(models.Model):
-    mass = models.PositiveIntegerField(blank = True, null = True)
+    mass = models.FloatField(blank = True, null = True)
     day = models.PositiveIntegerField(blank = True, null=True)
     month = models.PositiveIntegerField(blank = True, null=True)
     year = models.PositiveIntegerField(blank = True, null=False)
